reworked the classes

+++ src/main/java/com/manuel/decadev/model/Item.java
@@ -4,9 +4,9 @@ public abstract class Item {
    protected String category = "";
     protected int manufacturingDate = 0;
     protected String manufacturer = "";
-    protected int price = 0;
+    protected double price = 0;
 
-public Item(int price, String productCategory){
+public Item(double price, String productCategory){
     this.category = productCategory;
     this.price = price;
 }
